chore: rename profile actions
