+++ packages/sshnp_gui/lib/src/presentation/widgets/home_screen_table/home_screen_table_actions.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:sshnoports/sshnp/sshnp.dart';
 import 'package:sshnp_gui/src/presentation/widgets/home_screen_table/custom_table_cell.dart';
-import 'package:sshnp_gui/src/presentation/widgets/home_screen_table/home_screen_delete.dart';
-import 'package:sshnp_gui/src/presentation/widgets/home_screen_table/home_screen_edit.dart';
-import 'package:sshnp_gui/src/presentation/widgets/home_screen_table/home_screen_run.dart';
+import 'package:sshnp_gui/src/presentation/widgets/profile/actions/profile_delete_action.dart';
+import 'package:sshnp_gui/src/presentation/widgets/profile/actions/profile_edit_action.dart';
+import 'package:sshnp_gui/src/presentation/widgets/profile/actions/profile_run_action.dart';
 
 class HomeScreenTableActions extends StatelessWidget {
   final AsyncValue<SSHNPParams> params;
