+++ packages/sshnp_gui/lib/src/presentation/widgets/profile/actions/profile_delete_action.dart
