+++ packages/sshnp_gui/lib/src/presentation/widgets/profile/actions/profile_edit_action.dart
